+++ src/clap_app.rs
@@ -12,9 +12,10 @@ use crate::client::CloudManagerClient;
 use crate::config::CloudManagerConfig;
 use crate::encryption::{decrypt, encrypt};
 use crate::logs::{download_log, tail_log};
-use crate::models::{
-    Domain, EnvironmentVariableServiceType, LogType, PipelineVariableServiceType, ServiceType,
-};
+use crate::models::domain::Domain;
+use crate::models::log::{LogType, ServiceType};
+use crate::models::variables::{EnvironmentVariableServiceType, PipelineVariableServiceType};
+
 use crate::variables::{
     get_env_vars, get_pipeline_vars, set_env_vars_from_file, set_pipeline_vars_from_file,
 };
