+++ src/environments.rs
@@ -1,6 +1,7 @@
 use crate::client::{AdobeConnector, CloudManagerClient};
 use crate::errors::throw_adobe_api_error;
-use crate::models::{Environment, EnvironmentsList, EnvironmentsResponse};
+use crate::models::environment::{Environment, EnvironmentsList};
+use crate::models::variables::EnvironmentsResponse;
 use crate::HOST_NAME;
 use reqwest::{Error, Method};
 use std::process;
