+++ src/variables.rs
@@ -2,10 +2,11 @@ use crate::client::{AdobeConnector, CloudManagerClient};
 use crate::encryption::decrypt;
 use crate::environments::get_environment;
 use crate::errors::throw_adobe_api_error;
-use crate::models::{
+use crate::models::config::YamlConfig;
+use crate::models::variables::{
     EnvironmentVariable, EnvironmentVariableServiceType, EnvironmentVariablesList,
     EnvironmentVariablesResponse, PipelineVariable, PipelineVariableServiceType,
-    PipelineVariablesList, PipelineVariablesResponse, VariableType, YamlConfig,
+    PipelineVariablesList, PipelineVariablesResponse, VariableType,
 };
 use crate::pipelines::get_pipeline;
 use crate::HOST_NAME;
@@ -230,10 +231,7 @@ pub async fn set_env_vars_from_file(
                     }
 
                     if dry_run {
-                        println!(
-                            "{:>8} --dry-run detected. Not performing any actions.",
-                            "⚠️",
-                        );
+                        println!("{:>8} --dry-run detected. Not performing any actions.", "⚠️",);
                     } else {
                         match set_env_vars(client, p.id, e.id, &vars_final).await {
                             Ok(status) => match status {
@@ -474,10 +472,7 @@ pub async fn set_pipeline_vars_from_file(
                     }
 
                     if dry_run {
-                        println!(
-                            "{:>8} --dry-run detected. Not performing any actions.",
-                            "⚠️",
-                        );
+                        println!("{:>8} --dry-run detected. Not performing any actions.", "⚠️",);
                     } else {
                         match set_pipeline_vars(client, p.id, l.id, &vars_final).await {
                             Ok(status) => match status {
